requirements.txt
streamlit
pyJWT
snowflake-snowpark-python
pandas
plotly
